무기 및 아이템이 빛 효과로 인해 그림자가 생기지 않도록 조치 & 유로에셋 가지고 있던 거 하나 더 추가
무기 교체 완료(해당 프리팹은 Prefabs/Final Weapon에 있습니다.) & DamageNumbersPro 추가 데미지를 뜨게 만들 수 있으면 만들 예정입니다. & 플레이어에게 뎀지 안들어가는거 확인 필요(왠지 이것도 브로드캐스팅으로 해야할거같아서 확인필요)

+++ Assets/Scripts/Game Scene Script/EnemyScripts/Boss.cs
@@ -8,8 +8,8 @@ using UnityEngine.UIElements;
 public class Boss : MonoBehaviour
 {
     public int Healing;
-    public int maxHealth;
-    public int currentHealth;
+    public float maxHealth;
+    public float currentHealth;
 
     public Transform target;
     public BoxCollider meleeArea;
@@ -257,9 +257,9 @@ public class Boss : MonoBehaviour
         draw.Play();
         anim.SetTrigger(DoBigShot);
 
-        int previousHealth = enemyHealth.currentHealth;
-        int potentialHealth = Mathf.Clamp(enemyHealth.currentHealth + Healing, 0, maxHealth);
-        int healedAmount = potentialHealth - previousHealth;
+        float previousHealth = enemyHealth.currentHealth;
+        float potentialHealth = Mathf.Clamp(enemyHealth.currentHealth + Healing, 0, maxHealth);
+        float healedAmount = potentialHealth - previousHealth;
 
         if (healedAmount > 0)
         {
